+++ src/jsontoolkit/src/json.cc
@@ -1,14 +1,9 @@
-#include "tokens.h"
-#include "utils.h"
+#include "parsers/parser.h"
 #include <jsontoolkit/json.h>
 
 #include <stdexcept> // std::domain_error
 #include <string>    // std::to_string
 #include <utility>   // std::in_place_type
-
-static const char *const JSON_NULL = "null";
-static const char *const JSON_TRUE = "true";
-static const char *const JSON_FALSE = "false";
 
 sourcemeta::jsontoolkit::JSON::JSON(const char *const document)
     : source{document} {}
@@ -43,24 +38,25 @@ sourcemeta::jsontoolkit::JSON::JSON(sourcemeta::jsontoolkit::String &value)
            std::make_shared<sourcemeta::jsontoolkit::String>(value)} {}
 
 sourcemeta::jsontoolkit::JSON::JSON(const bool value)
-    : source{value ? JSON_TRUE : JSON_FALSE},
+    : source{value ? sourcemeta::jsontoolkit::parser::JSON_TRUE
+                   : sourcemeta::jsontoolkit::parser::JSON_FALSE},
       must_parse{false}, data{std::in_place_type<bool>, value} {}
 
 sourcemeta::jsontoolkit::JSON::JSON(const std::nullptr_t)
-    : source{JSON_NULL},
+    : source{sourcemeta::jsontoolkit::parser::JSON_NULL},
       must_parse{false}, data{std::in_place_type<std::nullptr_t>, nullptr} {}
 
 auto sourcemeta::jsontoolkit::JSON::parse() -> sourcemeta::jsontoolkit::JSON & {
   if (this->must_parse) {
     const std::string_view document =
-        sourcemeta::jsontoolkit::trim(this->source);
+        sourcemeta::jsontoolkit::parser::trim(this->source);
     std::variant<std::int64_t, double> number_result;
 
     switch (document.front()) {
-    case sourcemeta::jsontoolkit::JSON_ARRAY_START:
+    case sourcemeta::jsontoolkit::parser::JSON_ARRAY_START:
       this->data = std::make_shared<sourcemeta::jsontoolkit::Array>(document);
       break;
-    case sourcemeta::jsontoolkit::JSON_STRING_QUOTE:
+    case sourcemeta::jsontoolkit::parser::JSON_STRING_QUOTE:
       this->data = std::make_shared<sourcemeta::jsontoolkit::String>(document);
       break;
 
@@ -68,18 +64,18 @@ auto sourcemeta::jsontoolkit::JSON::parse() -> sourcemeta::jsontoolkit::JSON & {
     // zero. It may have a preceding minus sign (U+002D).
     // See
     // https://www.ecma-international.org/wp-content/uploads/ECMA-404_2nd_edition_december_2017.pdf
-    case sourcemeta::jsontoolkit::JSON_MINUS:
-    case sourcemeta::jsontoolkit::JSON_ZERO:
-    case sourcemeta::jsontoolkit::JSON_ONE:
-    case sourcemeta::jsontoolkit::JSON_TWO:
-    case sourcemeta::jsontoolkit::JSON_THREE:
-    case sourcemeta::jsontoolkit::JSON_FOUR:
-    case sourcemeta::jsontoolkit::JSON_FIVE:
-    case sourcemeta::jsontoolkit::JSON_SIX:
-    case sourcemeta::jsontoolkit::JSON_SEVEN:
-    case sourcemeta::jsontoolkit::JSON_EIGHT:
-    case sourcemeta::jsontoolkit::JSON_NINE:
-      number_result = sourcemeta::jsontoolkit::parse_number(document);
+    case sourcemeta::jsontoolkit::parser::JSON_MINUS:
+    case sourcemeta::jsontoolkit::parser::JSON_ZERO:
+    case sourcemeta::jsontoolkit::parser::JSON_ONE:
+    case sourcemeta::jsontoolkit::parser::JSON_TWO:
+    case sourcemeta::jsontoolkit::parser::JSON_THREE:
+    case sourcemeta::jsontoolkit::parser::JSON_FOUR:
+    case sourcemeta::jsontoolkit::parser::JSON_FIVE:
+    case sourcemeta::jsontoolkit::parser::JSON_SIX:
+    case sourcemeta::jsontoolkit::parser::JSON_SEVEN:
+    case sourcemeta::jsontoolkit::parser::JSON_EIGHT:
+    case sourcemeta::jsontoolkit::parser::JSON_NINE:
+      number_result = sourcemeta::jsontoolkit::parser::number(document);
       if (std::holds_alternative<std::int64_t>(number_result)) {
         this->data = std::get<std::int64_t>(number_result);
       } else {
