+++ src/jsontoolkit/src/json_array.cc
@@ -1,11 +1,7 @@
-#include <stdexcept> // std::domain_error
-#include <vector>    // std::vector
-
 #include <jsontoolkit/json.h>
 #include <jsontoolkit/json_array.h>
 
-#include "tokens.h"
-#include "utils.h"
+#include "parsers/parser.h"
 
 template <typename Wrapper, typename Backend>
 sourcemeta::jsontoolkit::GenericArray<Wrapper, Backend>::GenericArray()
@@ -33,109 +29,11 @@ auto sourcemeta::jsontoolkit::GenericArray<Wrapper, Backend>::size() ->
 template <typename Wrapper, typename Backend>
 auto sourcemeta::jsontoolkit::GenericArray<Wrapper, Backend>::parse()
     -> sourcemeta::jsontoolkit::GenericArray<Wrapper, Backend> & {
-  if (!this->must_parse) {
-    return *this;
+  if (this->must_parse) {
+    sourcemeta::jsontoolkit::parser::array<Wrapper>(this->source, this->data);
+    this->must_parse = false;
   }
 
-  const std::string_view document = sourcemeta::jsontoolkit::trim(this->source);
-  if (document.front() != sourcemeta::jsontoolkit::JSON_ARRAY_START ||
-      document.back() != sourcemeta::jsontoolkit::JSON_ARRAY_END) {
-    throw std::domain_error("Invalid array");
-  }
-
-  const std::string_view::size_type size = document.size();
-  std::string_view::size_type element_start_index = 0;
-  std::string_view::size_type level = 0;
-  bool is_string = false;
-
-  for (std::string_view::size_type index = 1; index < size - 1; index++) {
-    std::string_view::const_reference character = document.at(index);
-    const bool is_last_character = index == size - 2;
-
-    switch (character) {
-    case sourcemeta::jsontoolkit::JSON_ARRAY_START:
-      if (is_string) {
-        break;
-      }
-
-      // The start of an array at level 0 is by definition a new element
-      if (level == 0) {
-        element_start_index = index;
-      }
-
-      level += 1;
-      break;
-    case sourcemeta::jsontoolkit::JSON_ARRAY_END:
-      if (is_string) {
-        break;
-      }
-
-      if (level == 0) {
-        throw std::domain_error("Unexpected right bracket");
-      }
-
-      level -= 1;
-
-      // Only push an element on a final right bracket
-      if (is_last_character && element_start_index > 0) {
-        this->data.push_back(Wrapper(document.substr(
-            element_start_index, index - element_start_index + 1)));
-        element_start_index = 0;
-      }
-
-      break;
-    case sourcemeta::jsontoolkit::JSON_ARRAY_SEPARATOR:
-      if (is_string) {
-        break;
-      }
-
-      if (element_start_index == 0) {
-        throw std::domain_error("Separator without a preceding element");
-      }
-
-      if (is_last_character) {
-        throw std::domain_error("Trailing comma");
-      }
-
-      if (level == 0) {
-        this->data.push_back(Wrapper(
-            document.substr(element_start_index, index - element_start_index)));
-        element_start_index = 0;
-      }
-
-      break;
-    default:
-      if (is_last_character && element_start_index > 0) {
-        this->data.push_back(Wrapper(document.substr(
-            element_start_index, index - element_start_index + 1)));
-        element_start_index = 0;
-      } else if (is_last_character && element_start_index == 0 &&
-                 !sourcemeta::jsontoolkit::is_blank(character)) {
-        this->data.push_back(Wrapper(document.substr(index, 1)));
-      } else if (!sourcemeta::jsontoolkit::is_blank(character) &&
-                 element_start_index == 0 && level == 0) {
-        element_start_index = index;
-      }
-
-      if (character == sourcemeta::jsontoolkit::JSON_STRING_QUOTE) {
-        if (is_string) {
-          is_string = false;
-        } else if (index == 0 ||
-                   document.at(index - 1) !=
-                       sourcemeta::jsontoolkit::JSON_STRING_ESCAPE_CHARACTER) {
-          is_string = true;
-        }
-      }
-
-      break;
-    }
-  }
-
-  if (level > 0) {
-    throw std::domain_error("Unbalanced array");
-  }
-
-  this->must_parse = false;
   return *this;
 }
