--- src/jsontoolkit/src/json_number.cc
@@ -1,86 +1,0 @@
-#include "tokens.h"
-#include "utils.h"
-#include <cmath> // std::modf
-#include <jsontoolkit/json_number.h>
-#include <stdexcept> // std::domain_error
-#include <string>    // std::stol, std::stod
-
-auto sourcemeta::jsontoolkit::parse_number(const std::string_view &input)
-    -> std::variant<std::int64_t, double> {
-  const std::string_view document = sourcemeta::jsontoolkit::trim(input);
-
-  /*
-   * Validate the input number and decide whether it is an integer or a double
-   */
-
-  // A JSON number starts with a digit or the minus sign
-  std::string_view::const_reference front = document.front();
-  if (front != sourcemeta::jsontoolkit::JSON_MINUS &&
-      !sourcemeta::jsontoolkit::is_digit(front)) {
-    throw std::domain_error("Invalid number");
-  }
-
-  // A JSON number ends with a digit
-  if (!sourcemeta::jsontoolkit::is_digit(document.back())) {
-    throw std::domain_error("Invalid number");
-  }
-
-  const std::string_view::size_type size = document.size();
-  if (front == sourcemeta::jsontoolkit::JSON_ZERO && size > 1) {
-    std::string_view::const_reference second = document.at(1);
-    if (second != sourcemeta::jsontoolkit::JSON_DECIMAL_POINT &&
-        second != sourcemeta::jsontoolkit::JSON_EXPONENT_UPPER &&
-        second != sourcemeta::jsontoolkit::JSON_EXPONENT_LOWER) {
-      throw std::domain_error("Invalid leading zero");
-    }
-  }
-
-  bool integer = true;
-  std::string_view::size_type exponential_index = 0;
-  for (std::string_view::size_type index = 1; index < size - 1; index++) {
-    std::string_view::const_reference character = document.at(index);
-    std::string_view::const_reference previous = document.at(index - 1);
-
-    if (!sourcemeta::jsontoolkit::is_digit(character) &&
-        character != sourcemeta::jsontoolkit::JSON_MINUS &&
-        character != sourcemeta::jsontoolkit::JSON_DECIMAL_POINT &&
-        character != sourcemeta::jsontoolkit::JSON_EXPONENT_LOWER &&
-        character != sourcemeta::jsontoolkit::JSON_EXPONENT_UPPER) {
-      throw std::domain_error("Invalid real number");
-    }
-
-    if (character == sourcemeta::jsontoolkit::JSON_MINUS &&
-        previous != sourcemeta::jsontoolkit::JSON_EXPONENT_UPPER &&
-        previous != sourcemeta::jsontoolkit::JSON_EXPONENT_LOWER) {
-      throw std::domain_error("Invalid minus sign");
-    }
-
-    if (character == sourcemeta::jsontoolkit::JSON_DECIMAL_POINT) {
-      if (!integer || previous == sourcemeta::jsontoolkit::JSON_MINUS) {
-        throw std::domain_error("Invalid real number");
-      }
-
-      integer = false;
-    } else if (character == sourcemeta::jsontoolkit::JSON_EXPONENT_LOWER ||
-               character == sourcemeta::jsontoolkit::JSON_EXPONENT_UPPER) {
-      if (!sourcemeta::jsontoolkit::is_digit(previous) ||
-          exponential_index != 0) {
-        throw std::domain_error("Invalid exponential number");
-      }
-
-      exponential_index = index;
-      integer = false;
-    }
-  }
-
-  /*
-   * Convert the value accordingly
-   */
-
-  // TODO: Can we avoid converting to std::string?
-  if (integer) {
-    return std::stol(std::string(document));
-  }
-
-  return std::stod(std::string(document));
-}
